// backend/src/modules/leads/domain/activity-type.enum.ts
export enum ActivityType {
  LeadCreated = 'lead_created',
  LeadUpdated = 'lead_updated',
  LeadDeleted = 'lead_deleted',
  StatusChanged = 'status_changed',
  NoteAdded = 'note_added',
  LeadViewed = 'lead_viewed',
  LeadRestored = 'lead_restored',
  FollowUpCreated = 'follow_up_created',
  FollowUpCompleted = 'follow_up_completed',
  FollowUpCancelled = 'follow_up_cancelled',
  LeadAssigned = 'lead_assigned',
  LeadReassigned = 'lead_reassigned',
  LeadUnassigned = 'lead_unassigned',
  LeadTransferred = 'lead_transferred',
  LeadAutoAssigned = 'lead_auto_assigned',
  AgentCapacityReached = 'agent_capacity_reached',
  AgentCapacityUpdated = 'agent_capacity_updated',
  LeadUnassignedNoCapacity = 'lead_unassigned_no_capacity',
  LeadCreatedViaWebhook = 'lead_created_via_webhook',
  LeadCreatedViaMeta = 'lead_created_via_meta',
  LeadCreatedViaTiktok = 'lead_created_via_tiktok',
  LeadCreatedViaSnapchat = 'lead_created_via_snapchat',
  DuplicateDetected = 'duplicate_detected',
  FirstContactRecorded = 'first_contact_recorded',
  LeadsExported = 'leads_exported',
  CallStarted = 'call_started',
  CallEnded = 'call_ended',
}